Device: Rework required features
Should have definitions in one place that are used for both checks and
enables to remove the possibility of forgetting the feature check when
adding a new one to enables.

FOR_EACH_PAIR is a bit dirty, but it should work on gcc, clang and msvc
(with compliant preprocessor enabled). However, it's still more appealing
than having the check code laid out per struct for all potential fields.

Runtime reflection (fn that takes in a list of the names of enabled
features, templated by feature struct type) sounds cleaner. Sadly, that
requires actually implementing reflection and for library types
specifically.