Clean up PCH, fixes gli glm conflict
glm/gtc/type_ptr.hpp defines make_vec4 which conflicts with gli's
implementation of it if included in pch. Most of the headers are
lightweight, so let's just remove them to avoid similar issues.

Also don't need gli in pch as it's only used in texture loading and it
can be confined in a single translation unit.