Reorganize sources
Separate include and source directories don't make much sense here as
this is not a library to be linked against. Navigation is simpler with
header and source side-by-side.

Add a simple folder structure to group the codebase more clearly. Aim is
that the folder structure acts as a kind of soft module boundary where
gfx doesn't depend on render/scene and scene doesn't depend on render.