Move file polling to an async task
Polling was hitting the previous lag limit pretty often and sometimes
got longer lag spikes. std::filesystem doesn't seem to have a monitor
API but this should be the next best thing. Changes don't have to be
reacted to immediately so the one frame extra latency is ok and laggy
polling doesn't affect the rest of the app.