Have all submodels of a model in the same BLAS
Should be more optimal for nice content as models are treated as a unit
anyway and the submodels share the transform. Of course, content with
large holes in model BBs will likely have worse perf now because there
is no splitting logic at load time and the TLAS will have more of overlap.